+++ alarms/models/overdrive_checkouts_alarms.py
@@ -1,14 +1,15 @@
 from alarms.alarm import Alarm
-from datetime import timedelta
+from datetime import timedelta, timezone
 from helpers.alarm_helper import (
     check_redshift_mismatch_alarm,
     check_no_records_found_alarm,
 )
 from helpers.overdrive_web_scraper import OverDriveWebScraper
 from helpers.query_helper import (
-    build_redshift_ebook_query,
-    build_redshift_overdrive_duplicate_platform_query,
-    build_redshift_overdrive_duplicate_checksum_query,
+    build_redshift_daily_ebook_query,
+    build_redshift_daily_overdrive_platform_query,
+    build_redshift_monthly_ebook_query,
+    build_redshift_monthly_overdrive_platform_query,
 )
 from nypl_py_utils.functions.log_helper import create_log
 
@@ -20,26 +21,68 @@ class OverDriveCheckoutsAlarms(Alarm):
             overdrive_credentials[0], overdrive_credentials[1]
         )
         self.logger = create_log("overdrive_checkouts_alarms")
-        self.date_to_test = self.yesterday_date - timedelta(days=4)
+        self.daily_test_date = self.yesterday_date - timedelta(days=4)
+        # This is to make monthly checks start date 36 days before the current day
+        self.monthly_test_start_date = self.yesterday_date - timedelta(days=35)
 
     def run_checks(self):
         self.logger.info("OverDrive Checkouts")
-        redshift_tables = ["patron_overdrive_checkouts", "title_overdrive_checkouts"]
 
         try:
-            overdrive_count = self.overdrive_client.get_count(self.date_to_test)
+            self.overdrive_client.overdrive_login()
+            overdrive_count = self.overdrive_client.get_count(
+                self.daily_test_date, self.daily_test_date
+            )
         except Exception as e:
             self.logger.error(f"Failed to scrape OverDrive Marketplace: {e}")
+            self.overdrive_client.quit_driver()
             return
 
-        self.logger.info(f"Checking OD record count from ({self.date_to_test})...")
+        self.logger.info(f"Checking OD record count from ({self.daily_test_date})...")
+        self._run_redshift_checks(overdrive_count, monthly_check=False)
+
+        # The 'weekday' method returns Thursday as 3 so this check will run every Friday
+        if self.yesterday_date.weekday() == 3:
+            try:
+                self.logger.info(
+                    f"Running weekly check for inconsistencies between overdrive and redshift for the dates between {self.monthly_test_start_date} - {self.daily_test_date}"
+                )
+                monthly_overdrive_count = self.overdrive_client.get_count(
+                    self.monthly_test_start_date, self.daily_test_date
+                )
+                self._run_redshift_checks(monthly_overdrive_count, monthly_check=True)
+            except Exception as e:
+                self.logger.error(f"Failed to scrape OverDrive Marketplace: {e}")
+                self.overdrive_client.quit_driver()
+                return
+
+        self.overdrive_client.quit_driver()
+        check_no_records_found_alarm(
+            logger=self.logger,
+            database_count=overdrive_count,
+            conditional=True,
+            database_type="OverDrive Marketplace",
+            date=self.daily_test_date,
+        )
+
+    def _run_redshift_checks(self, overdrive_count, monthly_check=False):
+        redshift_tables = ["patron_overdrive_checkouts", "title_overdrive_checkouts"]
         for redshift_table in redshift_tables:
             table = redshift_table + self.redshift_suffix
-            redshift_query = build_redshift_ebook_query(table, self.date_to_test)
+            if monthly_check:
+                redshift_query = build_redshift_monthly_ebook_query(
+                    table, self.monthly_test_start_date, self.daily_test_date
+                )
+            else:
+                redshift_query = build_redshift_daily_ebook_query(
+                    table, self.daily_test_date
+                )
             redshift_count = self.get_record_count(self.redshift_client, redshift_query)
 
             if "patron_overdrive_checkouts" in table:
-                redshift_count = self._adjust_redshift_count(table, redshift_count)
+                redshift_count = self._adjust_redshift_count(
+                    table, redshift_count, monthly_check
+                )
 
             check_redshift_mismatch_alarm(
                 logger=self.logger,
@@ -49,15 +92,9 @@ class OverDriveCheckoutsAlarms(Alarm):
                 redshift_count=redshift_count,
             )
 
-        check_no_records_found_alarm(
-            logger=self.logger,
-            database_count=overdrive_count,
-            conditional=True,
-            database_type="OverDrive Marketplace",
-            date=self.date_to_test,
-        )
-
-    def _adjust_redshift_count(self, redshift_table, initial_redshift_count):
+    def _adjust_redshift_count(
+        self, redshift_table, initial_redshift_count, monthly_check=False
+    ):
         """
         In OverDrive, when users download titles through different platforms, these
         transactions are all counted as one row. This is not the case for the
@@ -66,26 +103,23 @@ class OverDriveCheckoutsAlarms(Alarm):
         """
         self.redshift_client.connect()
 
-        duplicate_checksums = self.redshift_client.execute_query(
-            build_redshift_overdrive_duplicate_checksum_query(
-                redshift_table, self.date_to_test
+        if monthly_check:
+            overdrive_duplicate_query = build_redshift_monthly_overdrive_platform_query(
+                redshift_table,
+                self.monthly_test_start_date,
+                self.daily_test_date,
             )
+        else:
+            overdrive_duplicate_query = build_redshift_daily_overdrive_platform_query(
+                redshift_table, self.daily_test_date
+            )
+
+        duplicate_result = self.redshift_client.execute_query(overdrive_duplicate_query)
+        duplicate_count = (
+            int(duplicate_result[0][0])
+            if duplicate_result and duplicate_result[0][0]
+            else 0
         )
 
-        if not duplicate_checksums:
-            # if the resulting tuple is empty, return original count
-            return initial_redshift_count
-
-        adjusted_redshift_count = initial_redshift_count
-        duplicate_checksums = [checksum for checksum in duplicate_checksums[0]]
-
-        for checksum in duplicate_checksums:
-            platform_types = self.redshift_client.execute_query(
-                build_redshift_overdrive_duplicate_platform_query(
-                    redshift_table, self.date_to_test, checksum
-                )
-            )
-            adjusted_redshift_count -= len(platform_types) - 1
-
         self.redshift_client.close_connection()
-        return adjusted_redshift_count
+        return initial_redshift_count - duplicate_count